Feat: Discoverable Credentials
