+++ frontend/src/api/types/authorize.ts
@@ -3,7 +3,7 @@ export type JwtTokenType = 'Bearer' | 'DPoP' | 'Id' | 'Refresh';
 
 export interface LoginRequest {
     /// Validation: `email`
-    email: string;
+    email?: string;
     /// Validation: max 256
     password?: string;
     pow: string;
@@ -24,6 +24,8 @@ export interface LoginRequest {
     /// RFC 8707 resource indicator forwarded from the authorization request.
     /// Validation: PATTERN_URI
     resource?: string;
+    /// Validation: max 64
+    resident_key_token?: string;
 }
 
 export interface LoginRefreshRequest {
