+++ frontend/src/i18n/common/nl.ts
@@ -117,6 +117,13 @@ export const I18nNl: I18n = {
             username: 'Gebruikersnaam',
             validFor: 'Wachtwoord geldig voor {{ secs }} seconden',
         },
+        passkeys: {
+            type: 'Passkey Type',
+            types: ['Default', 'Resident Key'],
+            rkWarning: `CAUTION: Even though a Resident Key can be used for a direct login without
+                the need to provide the E-Mail address, an additional storage slot is necessary
+                for each application. You need to make sure there is enough space left.`,
+        },
         passwordConfirm: 'Wachtwoord bevestigen',
         passwordCurr: 'Huidig wachtwoord',
         passwordCurrReq: 'Huidig wachtwoord is vereist',
@@ -153,6 +160,8 @@ export const I18nNl: I18n = {
         userEnabled: 'Gebruiker ingeschakeld',
         userExpiry: 'Gebruiker vervalt',
         userVerifiedTooltip: 'Beveiligd met vingerafdruk of PIN',
+        residentKeyTooltip: `This is a Resident Key. You can use it for a direct login without
+            providing an E-Mail.`,
         webIdDesc: `U kunt de velden instellen die zichtbaar zijn via uw WebID.\nDit is een functie
             die door sommige netwerken wordt gebruikt voor gedecentraliseerde logins. Als u niet
             \nweet wat het is, heeft u het waarschijnlijk niet nodig.`,
