+++ frontend/src/i18n/common/ru.ts
@@ -117,6 +117,14 @@ export const I18nRu: I18n = {
             username: 'Имя пользователя',
             validFor: 'Пароль действителен {{ secs }} секунд',
         },
+
+        passkeys: {
+            type: 'Passkey Type',
+            types: ['Default', 'Resident Key'],
+            rkWarning: `CAUTION: Even though a Resident Key can be used for a direct login without
+                the need to provide the E-Mail address, an additional storage slot is necessary
+                for each application. You need to make sure there is enough space left.`,
+        },
         passwordConfirm: 'Подтверждение пароля',
         passwordCurr: 'Текущий пароль',
         passwordCurrReq: 'Необходимо ввести текущий пароль',
@@ -152,6 +160,8 @@ export const I18nRu: I18n = {
         userEnabled: 'Пользователь активен',
         userExpiry: 'Срок действия пользователя',
         userVerifiedTooltip: 'Защищено отпечатком пальца или ПИН-кодом',
+        residentKeyTooltip: `This is a Resident Key. You can use it for a direct login without
+            providing an E-Mail.`,
         webIdDesc: `Вы можете настроить поля, которые будут доступны через ваш WebID.\nЭто
             функция, используемая некоторыми сетями для децентрализованного входа. Если вы не знаете, что это,
             \nвам, скорее всего, это не нужно.`,
