+++ frontend/src/i18n/common/uk.ts
@@ -119,6 +119,13 @@ export const I18nUk: I18n = {
             username: "Ім'я користувача",
             validFor: 'Пароль дійсний {{ secs }} секунд',
         },
+        passkeys: {
+            type: 'Passkey Type',
+            types: ['Default', 'Resident Key'],
+            rkWarning: `CAUTION: Even though a Resident Key can be used for a direct login without
+                the need to provide the E-Mail address, an additional storage slot is necessary
+                for each application. You need to make sure there is enough space left.`,
+        },
         passwordConfirm: 'Підтвердіть пароль',
         passwordCurr: 'Поточний пароль',
         passwordCurrReq: "Поточний пароль обов'язковий",
@@ -155,6 +162,8 @@ export const I18nUk: I18n = {
         userEnabled: 'Активовано',
         userExpiry: 'Термін дії',
         userVerifiedTooltip: 'Захищено відбитком пальця або PIN-кодом',
+        residentKeyTooltip: `This is a Resident Key. You can use it for a direct login without
+            providing an E-Mail.`,
         webIdDesc: `Ви можете налаштувати поля, які будуть доступні через ваш WebID.\nЦя функція
             використовується деякими мережами для децентралізованого входу. Якщо ви не знаєте, що
             це,\nвам це, швидше за все, не потрібно.`,
