+++ frontend/src/lib/UserPasskey.svelte
@@ -5,6 +5,7 @@
     import Tooltip from '$lib5/Tooltip.svelte';
     import Button from '$lib5/button/Button.svelte';
     import { useI18n } from '$state/i18n.svelte';
+    import IconKey from '$icons/IconKey.svelte';
 
     let {
         passkey,
@@ -31,6 +32,13 @@
                 <Tooltip text={t.account.userVerifiedTooltip}>
                     <div style:margin-bottom="-.25rem">
                         <IconFingerprint width="1rem" color="hsl(var(--accent))" />
+                    </div>
+                </Tooltip>
+            {/if}
+            {#if passkey.resident_key}
+                <Tooltip text={t.account.residentKeyTooltip}>
+                    <div style:margin-bottom="-.25rem">
+                        <IconKey width="1rem" color="hsl(var(--accent))" />
                     </div>
                 </Tooltip>
             {/if}
