+++ src/api/src/openapi.rs
@@ -266,9 +266,9 @@ use utoipa::{OpenApi, openapi};
             entity::password::PasswordHashTimes,
             entity::roles::Role,
             entity::scopes::Scope,
-            entity::webauthn::WebauthnAdditionalData,
-            entity::webauthn::WebauthnLoginReq,
-            entity::webauthn::WebauthnServiceReq,
+            entity::webauthn::auth_data::WebauthnAdditionalData,
+            entity::webauthn::auth_req::WebauthnLoginReq,
+            entity::webauthn::auth_req::WebauthnServiceReq,
             entity::well_known::WellKnown,
 
             AccessGroup,
@@ -413,6 +413,8 @@ use utoipa::{OpenApi, openapi};
             OtpLoginResponse,
             OtpAuthStartResponse,
             OtpLoginFinishResponse,
+            PasskeyResponse,
+            ResidentKeyToken,
             PamGetentResponse,
             PamGroupResponse,
             PamGroupHostsCountResponse,
