+++ src/data/src/entity/users.rs
@@ -17,7 +17,8 @@ use crate::entity::theme::ThemeCssFull;
 use crate::entity::tos::ToS;
 use crate::entity::tos_user_accept::ToSUserAccept;
 use crate::entity::users_values::UserValues;
-use crate::entity::webauthn::{PasskeyEntity, WebauthnServiceReq};
+use crate::entity::webauthn::auth_req::WebauthnServiceReq;
+use crate::entity::webauthn::passkey::PasskeyEntity;
 use crate::events::event::Event;
 use crate::html::templates::{HtmlTemplate, UserEmailChangeConfirmHtml};
 use crate::language::Language;
