+++ src/data/src/entity/webauthn/auth_req.rs
@@ -1,0 +1,89 @@
+use crate::database::{Cache, DB};
+use crate::rauthy_config::RauthyConfig;
+use rauthy_common::utils::get_rand;
+use rauthy_error::{ErrorResponse, ErrorResponseType};
+use serde::{Deserialize, Serialize};
+use utoipa::ToSchema;
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, ToSchema)]
+pub struct WebauthnLoginReq {
+    pub code: String,
+    pub user_id: String,
+    pub header_loc: String,
+    pub header_origin: Option<String>,
+    pub tos_await_data: Option<WebauthnToSAwaitData>,
+    pub needs_user_update: bool,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, ToSchema)]
+pub struct WebauthnToSAwaitData {
+    pub auth_code: String,
+    pub auth_code_lifetime: i32,
+}
+
+// CRUD
+impl WebauthnLoginReq {
+    pub async fn delete(&self) -> Result<(), ErrorResponse> {
+        DB::hql().delete(Cache::Webauthn, self.code.clone()).await?;
+        Ok(())
+    }
+
+    pub async fn find_remove(code: String) -> Result<Self, ErrorResponse> {
+        let res: Option<Self> = DB::hql().get_remove(Cache::Webauthn, code).await?;
+        match res {
+            None => Err(ErrorResponse::new(
+                ErrorResponseType::NotFound,
+                "Webauthn Login Request Data not found",
+            )),
+            Some(res) => Ok(res),
+        }
+    }
+
+    pub async fn save(&self) -> Result<(), ErrorResponse> {
+        let ttl = Some(RauthyConfig::get().vars.webauthn.data_exp as i64);
+        DB::hql()
+            .put(Cache::Webauthn, self.code.clone(), self, ttl)
+            .await?;
+        Ok(())
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, ToSchema)]
+pub struct WebauthnServiceReq {
+    pub code: String,
+    pub user_id: String,
+}
+
+// CRUD
+impl WebauthnServiceReq {
+    pub fn new(user_id: String) -> Self {
+        Self {
+            code: get_rand(48),
+            user_id,
+        }
+    }
+
+    pub async fn delete(&self) -> Result<(), ErrorResponse> {
+        DB::hql().delete(Cache::Webauthn, self.code.clone()).await?;
+        Ok(())
+    }
+
+    pub async fn find(code: String) -> Result<Self, ErrorResponse> {
+        let res = DB::hql().get(Cache::Webauthn, code).await?;
+        match res {
+            None => Err(ErrorResponse::new(
+                ErrorResponseType::NotFound,
+                "Webauthn Service Request Data not found",
+            )),
+            Some(res) => Ok(res),
+        }
+    }
+
+    pub async fn save(&self) -> Result<(), ErrorResponse> {
+        let ttl = Some(RauthyConfig::get().vars.webauthn.data_exp as i64);
+        DB::hql()
+            .put(Cache::Webauthn, self.code.clone(), self, ttl)
+            .await?;
+        Ok(())
+    }
+}
