+++ src/data/src/entity/webauthn/mod.rs
@@ -1,0 +1,7 @@
+pub mod auth_data;
+pub mod auth_req;
+pub mod authenticate;
+pub mod authenticate_rk;
+pub mod passkey;
+pub mod register;
+pub mod rk_token;
